Division de las clases, por archivos
